Add streaming voice command support (#35)
* Add streaming voice command support

Support sending voice commands with the RemoteVoice* messages.
Audio data needs to be 16-bit PCM, mono, 8000 Hz.
The Android TV Remote service appears to support audio configuration in
the RemoteVoiceBegin message, but no information could be found so far.

Add voice features with PyAudio to the demo application to easily test
the voice command feature:
- Record and stream voice from the default audio input
- Record a playback a local WAV file
- Send a pre-recorded WAV file as a voice command

No new dependencies are required for the library, only for the demo app.

* fixup! Add streaming voice command support

PR feedback